+++ squallmodels/v1/golang/enforcer.go
@@ -73,7 +73,7 @@ func (o EnforcersList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o EnforcersList) Version() float64 {
+func (o EnforcersList) Version() int {
 
 	return 1.0
 }
