+++ squallmodels/v1/golang/fileaccesspolicy.go
@@ -42,7 +42,7 @@ func (o FileAccessPoliciesList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o FileAccessPoliciesList) Version() float64 {
+func (o FileAccessPoliciesList) Version() int {
 
 	return 1.0
 }
