+++ squallmodels/v1/golang/poke.go
@@ -38,7 +38,7 @@ func (o PokesList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o PokesList) Version() float64 {
+func (o PokesList) Version() int {
 
 	return 1.0
 }
