+++ squallmodels/v1/golang/policyrule.go
@@ -40,7 +40,7 @@ func (o PolicyRulesList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o PolicyRulesList) Version() float64 {
+func (o PolicyRulesList) Version() int {
 
 	return 1.0
 }
