+++ squallmodels/v1/golang/processingunit.go
@@ -76,7 +76,7 @@ func (o ProcessingUnitsList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o ProcessingUnitsList) Version() float64 {
+func (o ProcessingUnitsList) Version() int {
 
 	return 1.0
 }
