+++ squallmodels/v1/golang/role.go
@@ -38,7 +38,7 @@ func (o RolesList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o RolesList) Version() float64 {
+func (o RolesList) Version() int {
 
 	return 1.0
 }
