+++ squallmodels/v1/golang/tabulation.go
@@ -38,7 +38,7 @@ func (o TabulationsList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o TabulationsList) Version() float64 {
+func (o TabulationsList) Version() int {
 
 	return 1.0
 }
