+++ vincemodels/v1/golang/plan.go
@@ -38,7 +38,7 @@ func (o PlansList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o PlansList) Version() float64 {
+func (o PlansList) Version() int {
 
 	return 1.0
 }
